# web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CUSA SSO API",
    "version": "1.4.0",
    "description": "Custom first-party SSO: Authorization Code + PKCE S256, opaque tokens, application-scoped roles. Not a full OAuth/OIDC provider; no discovery, ID token, refresh token or dynamic registration. Examples are placeholders; never paste production secrets into browser documentation."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this documentation; production uses HTTPS"
    }
  ],
...
  "paths": {
    "/api/sso/authorize": {
      "get": {
        "operationId": "authorize",
        "summary": "เริ่มเข้าสู่ระบบ Service",
        "description": "Browser navigation หลัง BFF สร้างและเก็บ state/verifier ใน server session; code มีอายุ 90 วินาที ใช้ครั้งเดียว ต้องมีสมาชิกและ Role ของ Service และอนุมัติรายการข้อมูลก่อนออก code; Service ที่เปิดคิวอาจส่งไป /waiting ก่อน Google/MFA ห้ามเพิ่มพารามิเตอร์นอกสัญญา คิวไม่แทนการตรวจสิทธิ์หรือ PKCE",
        "tags": [
          "Authorization"
        ],
        "security": [],
        "parameters": [
          {
            "name": "response_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "code"
              ]
            },
            "description": "Authorization Code grant เท่านั้น"
          },
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Application ID จากหน้า Applications ไม่ใช่ Google Client ID"
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri",
              "maxLength": 2048
            },
            "description": "Callback ต้องตรงกับที่ลงทะเบียนทุกตัวอักษร"
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{32,128}$"
            },
            "description": "CSPRNG อย่างน้อย 32 bytes ผูกกับ BFF session และใช้ครั้งเดียว"
          },
          {
            "name": "code_challenge",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Opaque credential: CSPRNG 32 bytes, base64url without padding",
              "pattern": "^[A-Za-z0-9_-]{43}$"
            },
            "description": "BASE64URL(SHA256(verifier)) ไม่มี = padding"
          },
          {
            "name": "code_challenge_method",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "S256"
              ]
            },
            "description": "ไม่รองรับ plain"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "identity:read profile email",
              "maxLength": 255
            },
            "description": "Space-separated requested claim scopes. Must include identity:read, be unique, and be permitted by the registered service policy. User may decline optional scopes."
          }
        ],
        "responses": {
          "303": {
            "description": "Redirect to login, service enrollment or consent as required; code only after explicit approval.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "400": {
            "description": "Validation / invalid grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid request parameters",
                  "code": "invalid_request"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid or expired API key",
                  "code": "invalid_client"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope / CORS / account access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "API key does not permit this operation",
                  "code": "insufficient_scope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in the permitted scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Resource not found",
                  "code": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Policy version or activity identifier conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Reload the current policy or use the correct event identifier",
                  "code": "CONFLICT"
                }
              }
            }
          },
          "429": {
            "description": "Quota exceeded; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service request quota exceeded",
                  "code": "RATE_LIMITED"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "เวลารอเป็นวินาที ห้าม retry รัว",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service unavailable",
                  "code": "AUDIT_UNAVAILABLE"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unable to process request",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        }
      }
    },
    "/api/sso/login-context": {
      "get": {
        "operationId": "loginContext",
        "summary": "อ่านบริบทหน้าเข้าสู่ระบบ",
        "description": "ใช้โดยหน้า Login ของ CUSA SSO; ผู้เชื่อมต่อทั่วไปเริ่มที่ authorize ได้เลย ไม่ออก token/code",
        "tags": [
          "Authorization"
        ],
        "security": [],
        "parameters": [
          {
            "name": "response_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "code"
              ]
            },
            "description": "Authorization Code grant เท่านั้น"
          },
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Application ID จากหน้า Applications ไม่ใช่ Google Client ID"
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri",
              "maxLength": 2048
            },
            "description": "Callback ต้องตรงกับที่ลงทะเบียนทุกตัวอักษร"
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{32,128}$"
            },
            "description": "CSPRNG อย่างน้อย 32 bytes ผูกกับ BFF session และใช้ครั้งเดียว"
          },
          {
            "name": "code_challenge",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Opaque credential: CSPRNG 32 bytes, base64url without padding",
              "pattern": "^[A-Za-z0-9_-]{43}$"
            },
            "description": "BASE64URL(SHA256(verifier)) ไม่มี = padding"
          },
          {
            "name": "code_challenge_method",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "S256"
              ]
            },
            "description": "ไม่รองรับ plain"
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "identity:read profile email",
              "maxLength": 255
            },
            "description": "Space-separated requested claim scopes. Must include identity:read, be unique, and be permitted by the registered service policy. User may decline optional scopes."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginContext"
                },
                "example": {
                  "application": {
                    "name": "Member Portal",
                    "origin": "https://portal.example.com"
                  },
                  "returnTo": "/api/sso/authorize?..."
                }
              }
            }
          },
          "400": {
            "description": "Validation / invalid grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid request parameters",
                  "code": "invalid_request"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid or expired API key",
                  "code": "invalid_client"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope / CORS / account access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "API key does not permit this operation",
                  "code": "insufficient_scope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in the permitted scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Resource not found",
                  "code": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Policy version or activity identifier conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Reload the current policy or use the correct event identifier",
                  "code": "CONFLICT"
                }
              }
            }
          },
          "429": {
            "description": "Quota exceeded; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service request quota exceeded",
                  "code": "RATE_LIMITED"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "เวลารอเป็นวินาที ห้าม retry รัว",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service unavailable",
                  "code": "AUDIT_UNAVAILABLE"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unable to process request",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sso/introspect": {
      "post": {
        "operationId": "introspectToken",
        "summary": "ตรวจ token ก่อน protected operation",
        "description": "ต้องมี token:introspect ตรวจได้เฉพาะ token ของ application เดียวกัน HTTP 200 ไม่เท่ากับ authenticated ต้องตรวจ active, aud, exp และ roles; identity cache สูงสุด 5 วินาที",
        "tags": [
          "Tokens"
        ],
        "security": [
          {
            "ApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntrospectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Active identity หรือ inactive ไม่มีข้อมูลผู้ใช้",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ActiveToken"
                    },
                    {
                      "$ref": "#/components/schemas/InactiveToken"
                    }
                  ]
                },
                "examples": {
                  "active": {
                    "value": {
                      "sub": "11111111-1111-4111-8111-111111111111",
                      "email": "member@example.com",
                      "name": "Example Member",
                      "given_name": "Example",
                      "family_name": "Member",
                      "department": "ฝ่ายสมาชิก",
                      "roles": [
                        "viewer",
                        "approver"
                      ],
                      "aud": "22222222-2222-4222-8222-222222222222",
                      "active": true,
                      "exp": 2000000000,
                      "scope": "identity:read profile email"
                    }
                  },
                  "inactive": {
                    "value": {
                      "active": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation / invalid grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid request parameters",
                  "code": "invalid_request"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid or expired API key",
                  "code": "invalid_client"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope / CORS / account access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "API key does not permit this operation",
                  "code": "insufficient_scope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in the permitted scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Resource not found",
                  "code": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Policy version or activity identifier conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Reload the current policy or use the correct event identifier",
                  "code": "CONFLICT"
                }
              }
            }
          },
          "429": {
            "description": "Quota exceeded; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service request quota exceeded",
                  "code": "RATE_LIMITED"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "เวลารอเป็นวินาที ห้าม retry รัว",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service unavailable",
                  "code": "AUDIT_UNAVAILABLE"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unable to process request",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "curl",
            "source": "curl --request POST \"$SSO_ORIGIN/api/sso/introspect\" \\\n  --header 'Content-Type: application/json' \\\n  --header \"X-API-Key: $SSO_API_KEY\" \\\n  --data '{\"token\":\"<ACCESS_TOKEN>\"}'"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Application-bound backend secret. Scope identity:read, token:introspect, token:revoke or member:activity as required."
      },
      "CookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-cusa_session",
        "description": "HttpOnly Secure cookie in production; cusa_session on localhost development"
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "ข้อความที่ปลอดภัยสำหรับแสดงผล"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code; case-sensitive"
          },
          "requestId": {
            "type": "string",
            "description": "Server-generated correlation UUID ถ้ามี",
            "format": "uuid"
          }
        }
      },
...
      "InactiveToken": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "const": false
          }
        },
        "required": [
          "active"
        ],
        "additionalProperties": false
      },
...
      "IntrospectionRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque access token",
            "minLength": 1,
            "maxLength": 512
          }
        },
        "required": [
          "token"
        ],
        "additionalProperties": false
      },
      "LoginContext": {
        "type": "object",
        "properties": {
          "application": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "ชื่อที่ลงทะเบียน"
              },
              "origin": {
                "type": "string",
                "description": "Origin จาก callback",
                "format": "uri"
              }
            },
            "required": [
              "name",
              "origin"
            ],
            "additionalProperties": false
          },
          "returnTo": {
            "type": "string",
            "description": "Internal authorize path สำหรับ View ของ CUSA SSO"
          },
          "registration": {
            "type": "string",
            "enum": [
              "closed",
              "invite",
              "open"
            ]
          }
        },
        "required": [
          "application",
          "returnTo"
        ],
        "additionalProperties": false
      },
...
    }
  }
}